Add dynamic data from history model to hisotry card view

+++ Card Memory Game/Model/HistoryModel.swift
@@ -5,7 +5,7 @@ import Foundation
 struct HistoryModel {
     var date: Date
     var difficulty: Int
-    var time: Double
+    var time: Int
     var lives: Int
     var moves: Int
 }
